requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.9.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
blinker==1.9.0
Brotli==1.1.0
bs4==0.0.2
CacheControl==0.14.2
cachetools==5.5.2
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
channels==4.3.1
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.3
cssselect2==0.8.0
daphne==4.2.1
Django==5.1.7
django-celery-beat==2.8.1
django-channels==0.7.0
django-cors-headers==4.7.0
django-filter==23.5
django-restframework==0.0.1
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
drf-spectacular==0.27.2
et_xmlfile==2.0.0
fcm-django==2.2.1
firebase-admin==6.6.0
Flask==3.1.1
flask-cors==5.0.1
fonttools==4.58.4
frozenlist==1.7.0
geographiclib==2.0
geopy==2.4.1
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-firestore
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.70.0
grpcio-status
gunicorn==23.0.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jwcrypto==1.5.6
kombu==5.5.4
MarkupSafe==3.0.2
msgpack==1.1.0
multidict==6.6.4
mysqlclient==2.2.7
numpy==2.3.0
oauthlib==3.2.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
paramiko==3.5.1
pillow==11.2.1
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pymongo==4.13.2
PyMuPDF==1.25.3
PyMySQL==1.1.1
PyNaCl==1.5.0
pyOpenSSL==25.0.0
pyparsing==3.2.3
PyPDF2==3.0.1
pyphen==0.17.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
reportlab==4.4.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
swapper==1.4.0
tinycss2==1.4.0
tinyhtml5==2.0.0
twilio==9.7.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.0rc1
tzdata==2025.2
Unidecode==1.3.8
uritemplate==4.2.0
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.1.3
xlrd==2.0.1
yarl==1.20.1
zope.interface==7.2
zopfli==0.2.3.post1
django-allauth==65.11.2
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
